refactor(LotteryGame at el): move LotteryGame class at el to lottery package
