+++ src/test/java/com/jaenyeong/mission02/lotto/domain/lottery/LotteryNumbersTest.java
@@ -1,11 +1,10 @@
-package com.jaenyeong.mission02.lotto.domain;
+package com.jaenyeong.mission02.lotto.domain.lottery;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Arrays;
 import java.util.Comparator;
@@ -16,15 +15,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@DisplayName("Lotto 복권의 한 게임에 해당하는 LotteryGame 클래스 테스트")
-class LotteryGameTest {
+@DisplayName("Lotto 게임에 복권 번호 리스트의 래핑 객체(일급 컬렉션) LotteryNumbers 클래스 테스트")
+class LotteryNumbersTest {
 
     @Test
-    @DisplayName("LotteryGame 객체 자동 생성 테스트")
-    void createLotteryGameAutomatically() {
-        final LotteryGame game = LotteryGame.ofAuto();
+    @DisplayName("LottoNumbers 객체 생성 테스트")
+    void createLotteryNumbersAutomatically() {
+        final LotteryNumbers lotteryNumbers = LotteryNumbers.ofAuto();
 
-        final List<Integer> lottery = game.getLotteryNumbers();
+        final List<Integer> lottery = lotteryNumbers.getLotteryNumbers();
 
         final int max = 6;
         final int start = 1;
@@ -38,11 +37,11 @@ class LotteryGameTest {
 
     @ParameterizedTest
     @MethodSource("givenValidNumberList")
-    @DisplayName("사용자가 직접 올바른 번호를 지정하는 경우 LotteryGame 객체 생성 테스트")
-    void createLotteryGameManuallyWhenGivenValidNumbers(final List<Integer> givenNumbers) {
-        final LotteryGame game = LotteryGame.ofManual(givenNumbers);
+    @DisplayName("사용자가 직접 올바른 번호를 지정하는 경우 LottoNumbers 객체 생성 테스트")
+    void createLotteryNumberManuallyWhenGivenValidNumbers(final List<Integer> givenNumbers) {
+        final LotteryNumbers lotteryNumbers = LotteryNumbers.ofManual(givenNumbers);
 
-        final List<Integer> lottery = game.getLotteryNumbers();
+        final List<Integer> lottery = lotteryNumbers.getLotteryNumbers();
         givenNumbers.sort(Comparator.naturalOrder());
 
         final int max = 6;
@@ -66,10 +65,10 @@ class LotteryGameTest {
 
     @ParameterizedTest
     @MethodSource("givenInvalidNumberList")
-    @DisplayName("사용자가 직접 올바르지 않은 번호를 지정하는 경우 LotteryGame 객체 생성 테스트")
-    void createLotteryGameManuallyWhenGivenInvalidNumbers(final List<Integer> givenNumbers) {
+    @DisplayName("사용자가 직접 올바르지 않은 번호를 지정하는 경우 LottoNumbers 객체 생성 테스트")
+    void createLotteryNumberManuallyWhenGivenInvalidNumbers(final List<Integer> givenNumbers) {
         assertThatThrownBy(() -> {
-            final LotteryGame ticket = LotteryGame.ofManual(givenNumbers);
+            final LotteryNumbers lotteryNumbers = LotteryNumbers.ofManual(givenNumbers);
         })
             .isInstanceOf(IllegalArgumentException.class);
     }
@@ -82,27 +81,5 @@ class LotteryGameTest {
             Arguments.of(Arrays.asList(10, 20, 30, 40, 15)),
             Arguments.of(Arrays.asList(40, 41))
         );
-    }
-
-    @ParameterizedTest
-    @MethodSource("givenValidNumberList")
-    @DisplayName("주어진 게임 번호와 현재 번호를 비교하는 테스트")
-    void matchWinningNumber(final List<Integer> givenNumbers) {
-        final LotteryGame game = LotteryGame.ofManual(givenNumbers);
-        final LotteryGame winningGame = LotteryGame.ofManual(Arrays.asList(1, 2, 3, 4, 5, 6));
-
-        final int missMatch = 0;
-        final int allRight = 6;
-
-        assertThat(winningGame.checkWinTheLottery(game)).isBetween(missMatch, allRight);
-    }
-
-    @ParameterizedTest
-    @ValueSource(ints = {1_000, 2_000, 100_000, 1_000_000})
-    @DisplayName("주어진 금액으로 구매 가능한 복권 티켓의 수를 확인하는 테스트")
-    void checkNumberAvailableForBuy(final int money) {
-        final int numberOfGame = LotteryGame.howManyBuyGame(money);
-
-        assertEquals(numberOfGame, money / LotteryGame.PRICE);
     }
 }
